requirements.txt
autopep8==1.5.7
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cssselect==1.1.0
Flask==2.0.1
gevent==21.8.0
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
pycodestyle==2.7.0
pycparser==2.20
pyquery==1.4.3
requests==2.26.0
schedule==1.1.0
soupsieve==2.2.1
urllib3==1.26.7
waitress==2.0.0
Werkzeug==2.0.1
pillow==7.1.2
pyupload